Add SpecificationSet.load_single_package
This implements the wrapper algorithm for ingesting a directory of YAML
files with specifications for a package, and resolving their
inheritance.

A resolution loop is used because we don't know the correct order
for processing documents. Instead we loop until all specifications are
resolved (or until there's a stalemate, and raise an exception).