Add SpecificationSet.load_metrics_package
This is the easiest way to get a SpecificationSet: it simply loads all
specification data available from a verification framework metric
definition package. By default this is
https://github.com/lsst/validate_metrics.

Things like "validate_metrics" are being call an LSST Verification
Framework metrics package, and that name is reflected in the method.

EUPS dependencies:

- Make utils required, now used in specset.py
- Make validate_metrics a required dependency since it's the default.
- Now using pex.exceptions because of getPackageDir().