Add SpecificationSet.subset method
This is a lovely way to winnow SpecificationSet to just the
collection of specifications useful in a certain context.

For example, a specification checking afterburner might want to get just
the specifications for a certain metric:

>>> spec_set.subset('validate_drp.PA1')
<a SpecificationSet instance>

Currently filters by either package name or metric name. Eventually we
could filter by a provenance query as well.