Convert BlobBase into Blob mapping container
This converts BlobBase into a Blob object. Blobs can be concretely
instantiated now, and aren't intended to be subclassed. This makes Blob
a data container (specifically, a container for Datums), and keeps this
data container separate from user code that calculates Datums. This
makes it easier to serialize/deserialize Blobs, and makes it easier to
use Blobs as a user as well.

This commit implements methods that make Blob work like a mapping.