Use properties to test Metric.unit type
The problem with make Metric.unit a plain attribute is that we can't
internally ensure that Metric.unit is in fact an astropy.units.Unit
instance. Using a property getter and setter lets us act as a gatekeeper
to ensure type consistency.

I've also added a setter to Metric.unit_str so that the unit can be set
given a Unit-compatible string. This is perhaps less surprising to a
user than converting a string assigned to Metric.unit into a Unit behind
the scenes.

I also changed Metric.unit_str to not provide an
'dimensionless_unscaled' metric. The problem was that
astropy.units.Unit('dimensionless_unscaled') does not work, so it'd be
surprising for users to have Metric.unit_str provide something that
can't be converted into a Unit.

Instead, I've changed __str__ to use 'dimensionless_unscaled' insteady
of an empty string. Same effect, but more correct I think.