Refactor MetricSet and Metric yaml parsing
The intention is to make MetricSet the single collection type for
Metrics, like SpecificationSet is for Specifications. Leveraging the
Name class as an internal key, it's not unreasonable to store metrics
from multiple package in the same collection.

This commit also unifies the API for obtaining metrics from a
verification framework metrics package with SpecificationSet.
There are load_metrics_package and load_single_package classmethods.

MetricSet is a mapping type, and like SpecificationSet, uses Name
instances as keys. string names are still accepted for access, but
internally converted into Name instances for logic checking.

The Metric class as a deserialize() method to match Specification.